Cleaned and moved a lot of stuff
I pretty much moved everything from the original repository that we
weren't using into `other-algorithms`, and make the top-level much
cleaner.

Insertion Sort has also been changed, and is a lot faster now. The
original implementation had a large amount of overhead and as a result,
was taking forever, even for an insertion sort.

`comparison.cpp` will not work with this commit, since I haven't updated
the filenames inside of it yet.